Tweak menu bar styling/organisation Fixes #22
